Add files via upload
This code implements a Convolutional Neural Network (CNN) to classify images from the CIFAR-10 dataset. The model is built using TensorFlow and Keras, and it consists of multiple convolutional and pooling layers, followed by dense layers for classification. The model is trained for 10 epochs with an Adam optimizer and sparse categorical cross-entropy loss.

Additionally, the code visualizes the training process by plotting the accuracy and loss for both training and validation sets, providing insight into the model's performance over the epochs.